applications/triangle-detector-server/requirements.txt
backports.functools-lru-cache==1.4
cycler==0.10.0
Cython==0.27.3
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
six==1.11.0
subprocess32==3.2.7
